Refactor test from Transaction to PaymentInfo
