+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/snap/paymentinfo/promo/PromoDetailsTest.java
@@ -1,4 +1,4 @@
-package com.midtrans.sdk.corekit.models.snap.transaction.promo;
+package com.midtrans.sdk.corekit.models.snap.paymentinfo.promo;
 
 import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.promo.Promo;
 import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.promo.PromoDetails;
